Update devise to 3.2.3
We need override most common views for devise as it does not support i18n very
good for them (lots of static text). Until it will be fixed on devise side we
need to have own customization for it.

Move also devise.rb initializer to the template as it require some modification
before using.

Change-Id: I4db802d49789d68fba1d36fe87dc527795f85afc