Fix issue with first day.
In our app we are using assumption that Sunday = 0, for example for fullcalendar
(make a note that for Fullcalendard 2.0 it will not be an issue anymore, but it
is still not released), but for the user we want to display Monday as the first
day of the week. For that purpose there are few helper methods to achieve that.

We are also linking user configuration about which days should be visible with
all parts of the system where the available_days are used.

Change-Id: I05a61c711b9026ca01e13c9521f0129c1c03e563