Remove cache for timetable_forms
Don't understand why but when this is cached the timetable_forms.js is loaded only
first time. When it will be cached next time page will not include that js file
probably due to that this view next time will be taken from cache where the js
file is not available since it was "rendered" directly on layout.

Change-Id: I190e208c87140031bbb7c8f526c24da4c7f36b26